+++ fractal_server/app/routes/admin/v2/resource.py
@@ -4,12 +4,15 @@ from fastapi import HTTPException
 from fastapi import Response
 from fastapi import status
 from pydantic import ValidationError
+from sqlmodel import func
 from sqlmodel import select
 
+from ._aux_functions import _check_resource_name
 from ._aux_functions import _get_resource_or_404
 from fractal_server.app.db import AsyncSession
 from fractal_server.app.db import get_async_db
 from fractal_server.app.models import UserOAuth
+from fractal_server.app.models.v2 import Profile
 from fractal_server.app.models.v2 import Resource
 from fractal_server.app.routes.auth import current_active_superuser
 from fractal_server.app.schemas.v2 import ResourceCreate
@@ -73,6 +76,8 @@ async def post_resource(
             ),
         )
 
+    await _check_resource_name(name=resource_create.name, db=db)
+
     # Handle non-unique resource names
     res = await db.execute(
         select(Resource).where(Resource.name == resource_create.name)
@@ -109,21 +114,8 @@ async def patch_resource(
     resource = await _get_resource_or_404(resource_id=resource_id, db=db)
 
     # Handle non-unique resource names
-    if (
-        resource_update.name is not None
-        and resource_update.name != resource.name
-    ):
-        res = await db.execute(
-            select(Resource).where(Resource.name == resource_update.name)
-        )
-        if res.scalars().one_or_none():
-            raise HTTPException(
-                status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
-                detail=(
-                    f"Resource name '{resource_update.name}' "
-                    "already in use."
-                ),
-            )
+    if resource_update.name and resource_update.name != resource.name:
+        await _check_resource_name(name=resource_update.name, db=db)
 
     # Prepare new db object
     for key, value in resource_update.model_dump(exclude_unset=True).items():
@@ -156,6 +148,24 @@ async def delete_resource(
     Delete single `Resource`.
     """
     resource = await _get_resource_or_404(resource_id=resource_id, db=db)
+
+    # Fail if at least one Profile is associated with the Resource.
+    res = await db.execute(
+        select(func.count(Profile.id)).where(
+            Profile.resource_id == resource_id
+        )
+    )
+    associated_profile_count = res.scalar()
+    if associated_profile_count > 0:
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
+            detail=(
+                f"Cannot delete Resource {resource_id} because it's associated"
+                f" with {associated_profile_count} Profiles."
+            ),
+        )
+
+    # Delete
     await db.delete(resource)
     await db.commit()
